# artifacts/api-server/src/routes/collectors.ts
import { Router, type IRouter } from "express";
import { db, collectorsTable, branchesTable, collectionsTable } from "@workspace/db";
import { eq, and, ilike, sql } from "drizzle-orm";

const router: IRouter = Router();

router.get("/collectors", async (req, res): Promise<void> => {
  const { branchId, search } = req.query;

  let rows = await db
    .select({ col: collectorsTable, branchName: branchesTable.name })
    .from(collectorsTable)
    .leftJoin(branchesTable, eq(collectorsTable.branchId, branchesTable.id))
    .orderBy(collectorsTable.name);

  if (branchId) rows = rows.filter((r) => r.col.branchId === parseInt(branchId as string, 10));
  if (search && typeof search === "string") {
    rows = rows.filter((r) => r.col.name.toLowerCase().includes(search.toLowerCase()) || r.col.mobile.includes(search as string));
  }

  const result = await Promise.all(
    rows.map(async (row) => {
      const [stats] = await db
        .select({ count: sql<number>`count(*)::int`, total: sql<string>`coalesce(sum(amount),0)` })
        .from(collectionsTable)
        .where(eq(collectionsTable.collectorId, row.col.id));
      return {
        ...row.col,
        branchName: row.branchName,
        totalCollections: stats?.count ?? 0,
        totalAmount: parseFloat(stats?.total ?? "0"),
        createdAt: row.col.createdAt.toISOString(),
      };
    })
  );

  res.json(result);
});

router.post("/collectors", async (req, res): Promise<void> => {
  const { name, mobile, email, branchId, status } = req.body;
  if (!name || !mobile || !branchId) {
    res.status(400).json({ error: "name, mobile, branchId required" });
    return;
  }
  const [col] = await db.insert(collectorsTable).values({ name, mobile, email, branchId, status: status ?? "active" }).returning();
  res.status(201).json({ ...col, totalCollections: 0, totalAmount: 0, createdAt: col.createdAt.toISOString() });
});

router.get("/collectors/:id", async (req, res): Promise<void> => {
  const id = parseInt(Array.isArray(req.params.id) ? req.params.id[0] : req.params.id, 10);
  const [row] = await db
    .select({ col: collectorsTable, branchName: branchesTable.name })
    .from(collectorsTable)
    .leftJoin(branchesTable, eq(collectorsTable.branchId, branchesTable.id))
    .where(eq(collectorsTable.id, id));
  if (!row) { res.status(404).json({ error: "Collector not found" }); return; }
  const [stats] = await db
    .select({ count: sql<number>`count(*)::int`, total: sql<string>`coalesce(sum(amount),0)` })
    .from(collectionsTable)
    .where(eq(collectionsTable.collectorId, id));
  res.json({ ...row.col, branchName: row.branchName, totalCollections: stats?.count ?? 0, totalAmount: parseFloat(stats?.total ?? "0"), createdAt: row.col.createdAt.toISOString() });
});

router.patch("/collectors/:id", async (req, res): Promise<void> => {
  const id = parseInt(Array.isArray(req.params.id) ? req.params.id[0] : req.params.id, 10);
  const { name, mobile, email, branchId, status } = req.body;
  const [col] = await db.update(collectorsTable).set({ name, mobile, email, branchId, status }).where(eq(collectorsTable.id, id)).returning();
  if (!col) { res.status(404).json({ error: "Collector not found" }); return; }
  res.json({ ...col, totalCollections: 0, totalAmount: 0, createdAt: col.createdAt.toISOString() });
});

router.delete("/collectors/:id", async (req, res): Promise<void> => {
  const id = parseInt(Array.isArray(req.params.id) ? req.params.id[0] : req.params.id, 10);
  await db.delete(collectorsTable).where(eq(collectorsTable.id, id));
  res.sendStatus(204);
});

router.get("/collectors/:id/performance", async (req, res): Promise<void> => {
  const id = parseInt(Array.isArray(req.params.id) ? req.params.id[0] : req.params.id, 10);
  const [all] = await db
    .select({ count: sql<number>`count(*)::int`, total: sql<string>`coalesce(sum(amount),0)` })
    .from(collectionsTable)
    .where(eq(collectionsTable.collectorId, id));

  const now = new Date();
  const monthStart = new Date(now.getFullYear(), now.getMonth(), 1);
  const [month] = await db
    .select({ count: sql<number>`count(*)::int`, total: sql<string>`coalesce(sum(amount),0)` })
    .from(collectionsTable)
    .where(and(eq(collectionsTable.collectorId, id), sql`collected_at >= ${monthStart.toISOString()}`));

  const totalCount = all?.count ?? 0;
  const totalAmount = parseFloat(all?.total ?? "0");

  res.json({
    collectorId: id,
    totalCollections: totalCount,
    totalAmount,
    successRate: totalCount > 0 ? 95 : 0,
    thisMonthAmount: parseFloat(month?.total ?? "0"),
    thisMonthCount: month?.count ?? 0,
    dailyTrend: [],
  });
});

export default router;