# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bissi / Committee / Chit Fund Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /collectors:
    get:
      operationId: listCollectors
      tags: [collectors]
      summary: List collectors
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of collectors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collector"
    post:
      operationId: createCollector
      tags: [collectors]
      summary: Create a collector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectorInput"
      responses:
        "201":
          description: Collector created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collector"
  /collectors/{id}:
    get:
      operationId: getCollector
      tags: [collectors]
      summary: Get collector by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Collector
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collector"
    patch:
      operationId: updateCollector
      tags: [collectors]
      summary: Update a collector
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectorUpdate"
      responses:
        "200":
          description: Collector updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collector"
    delete:
      operationId: deleteCollector
      tags: [collectors]
      summary: Delete a collector
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /collectors/{id}/performance:
    get:
      operationId: getCollectorPerformance
      tags: [collectors]
      summary: Get collector performance summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Collector performance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectorPerformance"

  # Committees
  /committees:
    get:
      operationId: listCommittees
      tags: [committees]
      summary: List committees/bissi
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of committees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Committee"
    post:
      operationId: createCommittee
      tags: [committees]
      summary: Create a committee/bissi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitteeInput"
      responses:
        "201":
          description: Committee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Committee"
  /committees/{id}:
    get:
      operationId: getCommittee
      tags: [committees]
      summary: Get committee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Committee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Committee"
    patch:
      operationId: updateCommittee
      tags: [committees]
      summary: Update a committee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitteeUpdate"
      responses:
        "200":
          description: Committee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Committee"
  /committees/{id}/members:
    get:
      operationId: listCommitteeMembers
      tags: [committees]
      summary: List committee members
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommitteeMember"
    post:
      operationId: addCommitteeMember
      tags: [committees]
      summary: Add member to committee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitteeMemberInput"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitteeMember"

  # Tokens
...
components:
  schemas:
...
    TrendPoint:
      type: object
      required: [date, amount, count]
      properties:
        date:
          type: string
        amount:
          type: number
        count:
          type: integer
...
    Collector:
      type: object
      required: [id, name, mobile, branchId, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        email:
          type: ["string", "null"]
        branchId:
          type: integer
        branchName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        totalCollections:
          type: integer
        totalAmount:
          type: number
        createdAt:
          type: string
    CollectorInput:
      type: object
      required: [name, mobile, branchId]
      properties:
        name:
          type: string
        mobile:
          type: string
        email:
          type: string
        branchId:
          type: integer
        status:
          type: string
          enum: [active, inactive]
    CollectorUpdate:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        email:
          type: string
        branchId:
          type: integer
        status:
          type: string
          enum: [active, inactive]
    CollectorPerformance:
      type: object
      required: [collectorId, totalCollections, totalAmount, successRate, thisMonthAmount, thisMonthCount]
      properties:
        collectorId:
          type: integer
        totalCollections:
          type: integer
        totalAmount:
          type: number
        successRate:
          type: number
        thisMonthAmount:
          type: number
        thisMonthCount:
          type: integer
        dailyTrend:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
    Committee:
      type: object
      required: [id, name, type, installmentAmount, memberLimit, status, branchId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [daily, weekly, monthly, festival, special]
        installmentAmount:
          type: number
        memberLimit:
          type: integer
        currentMembers:
          type: integer
        drawDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, completed, cancelled]
        branchId:
          type: integer
        branchName:
          type: ["string", "null"]
        totalCollected:
          type: number
        createdAt:
          type: string
    CommitteeInput:
      type: object
      required: [name, type, installmentAmount, memberLimit, branchId]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [daily, weekly, monthly, festival, special]
        installmentAmount:
          type: number
        memberLimit:
          type: integer
        drawDate:
          type: string
        branchId:
          type: integer
        status:
          type: string
          enum: [active, completed, cancelled]
    CommitteeUpdate:
      type: object
      properties:
        name:
          type: string
        installmentAmount:
          type: number
        memberLimit:
          type: integer
        drawDate:
          type: string
        status:
          type: string
          enum: [active, completed, cancelled]
    CommitteeMember:
      type: object
      required: [id, committeeId, customerId, customerName, tokenNumber, joinedAt]
      properties:
        id:
          type: integer
        committeeId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        customerMobile:
          type: ["string", "null"]
        tokenNumber:
          type: string
        status:
          type: string
        joinedAt:
          type: string
    CommitteeMemberInput:
      type: object
      required: [customerId]
      properties:
        customerId:
          type: integer
        tokenNumber:
          type: string
...