# lib/api-zod/src/generated/types/committeeUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bissi / Committee / Chit Fund Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CommitteeUpdateStatus } from './committeeUpdateStatus';

export interface CommitteeUpdate {
  name?: string;
  installmentAmount?: number;
  memberLimit?: number;
  drawDate?: string;
  status?: CommitteeUpdateStatus;
}